% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/An_Functions.R
\name{assign_promiscuous_samples}
\alias{assign_promiscuous_samples}
\title{assign_promiscuous_samples}
\usage{
assign_promiscuous_samples(out_one_D, Dis_Comp_Data)
}
\arguments{
\item{out_one_D}{Output from one_D_Mapper function.}

\item{Dis_Comp_Data}{Disease component matrix.}
}
\value{

}
\description{
Assigns promiscuous samples to unique nodes.
}
\examples{
\dontrun{
assign_promiscuous_samples(out_one_D,Dis_Comp_Data)
}
}
