% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{aux_function_select_top_diff}
\alias{aux_function_select_top_diff}
\title{aux_function_select_top_diff}
\usage{
aux_function_select_top_diff(x, n_genes)
}
\arguments{
\item{x}{res_diff_exp object}

\item{n_genes}{number of top genes.}
}
\value{

}
\description{
Axuliar function to select_top_diff_genes_groups
}
\examples{
\dontrun{
aux_function_select_top_diff(x,n_genes)
}
}
