% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{axuiliar_function_to_mean}
\alias{axuiliar_function_to_mean}
\title{axuiliar_function_to_mean}
\usage{
axuiliar_function_to_mean(x, y)
}
\arguments{
\item{x}{vector of numeric values.}

\item{y}{vector indicating the group to which each numeric value belong.}
}
\value{

}
\description{
Auxiliar function that computes the mean of a vector based on a group.
}
\examples{
\dontrun{
axuiliar_function_to_mean(x,y)
}
}
