% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{block_wise_mod}
\alias{block_wise_mod}
\title{block_wise_mod}
\usage{
block_wise_mod(list_exp, powers)
}
\arguments{
\item{list_exp}{list containing gene expressoin datasets.}

\item{powers}{power derived from the scale-free topology fit analysis.}
}
\value{

}
\description{
Computes blockwiseConsensusModules
}
\examples{
\dontrun{
block_wise_mod(list_exp, powers)
}
}
