% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{clust_all_levels}
\alias{clust_all_levels}
\title{clust_all_levels}
\usage{
clust_all_levels(
  Dis_Est_Mod,
  samp_in_lev,
  distance_type = c("cor", "euclidean"),
  optimal_clust_mode = c("standard", "silhouette"),
  n_bins_clust = 10
)
}
\arguments{
\item{Dis_Est_Mod}{Disease estate model data for the complete dataset.}

\item{samp_in_lev}{A list including the samples placed at each level. It is the output of the samples_in_levels function.}

\item{distance_type}{The distance type employed.}

\item{optimal_clust_mode}{The optimal cluster method selection.}

\item{n_bins_clust}{The number of bines for histogram threshold selection. Used in the standard method for optimal number of clusters selection.}
}
\value{
Returns a list including the clusters found at each level.
}
\description{
clust_all_levels
}
\examples{
\dontrun{
samples_in_levels(Dis_Est_Mod,samp_in_lev)}
}
