% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{clust_lev}
\alias{clust_lev}
\title{clust_lev}
\usage{
clust_lev(
  dis_est_mod_lev,
  distance_type = "cor",
  clust_type = "hierarchical",
  linkage_type = "single",
  optimal_clust_mode = "standard",
  n_bins_clust = 10,
  level_name = "level_1"
)
}
\arguments{
\item{dis_est_mod_lev}{Disease estate data for the whole datasets and the selected genes.}

\item{distance_type}{Type of distance , correlation, euclidean...}

\item{optimal_clust_mode}{Method for selection optimal number of clusters. For the moment only the standard method has been implemented.}

\item{n_bins_clust}{Number of bins in order to select the optimal threshold for number of cluster computations.}

\item{level_name}{Name of the studied level.}
}
\value{
Returns the samples included in each cluster for the specific level analyzed.
}
\description{
Get clusters for a particular data level.
}
\examples{
\dontrun{
clust_lev(dis_est_mod_lev,distance_type,optimal_clust_mode,n_bins_clust,level_name)}
}
