% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{compute_node_adjacency}
\alias{compute_node_adjacency}
\title{compute_node_adjacency}
\usage{
compute_node_adjacency(nodes_list)
}
\arguments{
\item{nodes_list}{A list with the sample content of each node.}
}
\value{
Returns a matrix object that stores a 1 if there are shared samples in two given nodos and a 0 otherwise.
}
\description{
Computes the adjacency matrix.
}
\examples{
\dontrun{
compute_node_adjacency(nodes_list)}
}
