% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{compute_power_tables}
\alias{compute_power_tables}
\title{compute_power_tables}
\usage{
compute_power_tables(list_datasets)
}
\arguments{
\item{list_datasets}{spitted list of expression matrices.}
}
\value{

}
\description{
Compute power tables for each dataset.
}
\examples{
\dontrun{
compute_power_tables(dis_st_mod,filt_vector,n_int,p)
}
}
