% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/An_Functions.R
\name{compute_u_mat_and_eigen}
\alias{compute_u_mat_and_eigen}
\title{compute_u_mat_and_eigen}
\usage{
compute_u_mat_and_eigen(dis_comp_node)
}
\arguments{
\item{dis_comp_node}{Subset of the disease component matrix.}
}
\value{

}
\description{
Computes the matrix U and the singular values for a subset of samples placed in a node.
}
\examples{
\dontrun{
compute_u_mat_and_eigen(dis_comp_node)
}
}
