% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{create_data_frame_from_multi}
\alias{create_data_frame_from_multi}
\title{create_data_frame_from_multi}
\usage{
create_data_frame_from_multi(vec_to_trans, row_names)
}
\arguments{
\item{vec_to_trans}{vector of covariate with more than 2 values.}

\item{row_names}{rownames of the original dataframe}
}
\value{

}
\description{
create_data_frame_from_multi
}
\examples{
\dontrun{
create_data_frame_from_multi(vec_to_trans,row_names)
}
}
