% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{create_multi_expressoin_object}
\alias{create_multi_expressoin_object}
\title{create_multi_expressoin_object}
\usage{
create_multi_expressoin_object(test_splitted)
}
\arguments{
\item{test_splitted}{test_splitted}
}
\value{

}
\description{
create_multi_expressoin_object
}
\examples{
\dontrun{
create_multi_expressoin_object(test_splitted)
}
}
