% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flat_Construction.R
\name{flatten_normal_tiss}
\alias{flatten_normal_tiss}
\title{F}
\usage{
flatten_normal_tiss(normal_tiss)
}
\arguments{
\item{normal_tissue_data}{}
}
\value{

}
\description{
F
}
\examples{
flatten_normal_tiss(normal_tissue_matrix)
}
