% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{format_results}
\alias{format_results}
\title{format_results}
\usage{
format_results(rsults)
}
\arguments{
\item{rsults}{results from the}
}
\value{

}
\description{
Formats the differential gene expression results from perform_wilcoxon_each to write into an excel file.
}
