% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{fun_to_app_prep}
\alias{fun_to_app_prep}
\title{fun_to_app_prep}
\usage{
fun_to_app_prep(x, n_genes)
}
\arguments{
\item{x}{A dataframe}

\item{n_genes}{Number of genes showing the highest differences in methylation.}
}
\value{

}
\description{
Auxiliar function for prepare_diff_meth_res
}
\examples{
\dontrun{
fun_to_app_prep(x,n_genes)
}
}
