% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Select_And_Surv.R
\name{gene_selection}
\alias{gene_selection}
\title{Gene selection based on variability}
\usage{
gene_selection(disease_component_tumors, percent = 0.85)
}
\arguments{
\item{disease_component_tumors}{A matrix with the disease component data of the entire dataset.}

\item{percent}{Percentile for gene selection.}
}
\value{
A vector containing the names of the selected genes.
}
\description{
Gene selection based on variability
}
\examples{
\dontrun{
gene_selection(disease_component_tumors,0.99)
}
}
