% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Select_And_Surv.R
\name{gene_selection_surv}
\alias{gene_selection_surv}
\title{gene_selection_surv}
\usage{
gene_selection_surv(
  D_Comp,
  p_Data,
  Status_Col_Name,
  Status_Value,
  Cox_All,
  n_top,
  type_sel = c("Top_Bot", "Abs")
)
}
\arguments{
\item{D_Comp}{Disease component object}

\item{p_Data}{Data.frame with the phenotype data.}

\item{Status_Col_Name}{Column for sample filtering.}

\item{Status_Value}{Value for the sample filtering co-variate.}

\item{Cox_All}{Output from the cox_all_genes function.}

\item{n_top}{Number of markers presenting the maximum values for the product.}

\item{type_sel}{Select top from bottom and top or from absolute value.}
}
\value{

}
\description{
Selects markers for mapper based on the product of standard deviation plus one times the Z score obtained thorugh fitting a cox proportional hazard model to the expression of each gene.
}
\examples{
\dontrun{
gene_selection_surv(D_Comp,p_Data,Status_Col_Name,Status_Value,Cox_All,n_top)
}
}
