% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/An_Functions.R
\name{generate_df_samp_to_node}
\alias{generate_df_samp_to_node}
\title{generate_df_samp_to_node}
\usage{
generate_df_samp_to_node(out_one_D)
}
\arguments{
\item{out_one_D}{Results object obtained from the}
}
\value{

}
\description{
Auxiliary function for the generation of the output data.frame.
}
\examples{
\dontrun{
generate_df_samp_to_node(out_one_D)
}
}
