% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/An_Functions.R
\name{generate_not_ov_clust}
\alias{generate_not_ov_clust}
\title{generate_not_ov_clust}
\usage{
generate_not_ov_clust(out_one_D, exp_filt)
}
\arguments{
\item{out_one_D}{Results object obtained from the}

\item{exp_filt}{filtered expression matrix.}
}
\value{

}
\description{
Auxiliary function for the computation of principal components.
}
\examples{
\dontrun{
generate_not_ov_clust(out_one_D,exp_filt)
}
}
