% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/An_Functions.R
\name{get_information_from_results}
\alias{get_information_from_results}
\title{get_information_from_results}
\usage{
get_information_from_results(res_mapper)
}
\arguments{
\item{res_mapper}{Result object from a mapper function.}
}
\value{

}
\description{
This functions retrieves information about the mapper results.
}
\examples{
\dontrun{
get_information_from_results(res_mapper)
}
}
