% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{get_intervals_One_D}
\alias{get_intervals_One_D}
\title{Extract intervals from filter function output values.}
\usage{
get_intervals_One_D(dis_st_mod, filt_vector, n_int, p)
}
\arguments{
\item{dis_st_mod}{Disease-state model matrix generated using the complete dataset.}

\item{filt_vector}{Vector obtained after applying the filtering function to the disease-state model data.}

\item{n_int}{Number of intervals to divide the filtering function in.}

\item{p}{Percentage of overlap between intervals.}
}
\value{
Returns the set of intervals for the filtering function.
}
\description{
This function constructs the intervals for a given filtering function.
}
\examples{
\dontrun{
get_intervals_One_D(dis_st_mod,filt_vector,n_int,p)}
}
