% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flat_Construction.R
\name{get_lambda}
\alias{get_lambda}
\title{Computes lambda}
\usage{
get_lambda(bet)
}
\arguments{
\item{bet}{numeric}
}
\value{
numeric
}
\description{
Computes the value of lambda as defined in: The Optimal Hard Thresholdfor Singular Values is \deqn{\sqrt(4/ 3)}
}
\examples{
get_lambda(0.3)
}
