% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{ggplotColours}
\alias{ggplotColours}
\title{Function to get colors like in ggplot2}
\usage{
ggplotColours(n = 6, h = c(0, 360) + 15)
}
\arguments{
\item{n}{number of colors.}

\item{h}{range.}
}
\value{

}
\description{
Function to get colors like in ggplot2
}
\examples{
\dontrun{
ggplotColours(n,h)}
}
