% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{levels_to_nodes}
\alias{levels_to_nodes}
\title{levels_to_nodes}
\usage{
levels_to_nodes(clust_all_levels_list)
}
\arguments{
\item{clust_all_levels_list}{A list obtained from the clust all levels function.}
}
\value{
A list including the sample content of each detected node.
}
\description{
Extract the nodes information based on the level clustering data.
}
\examples{
\dontrun{
levels_to_nodes(clust_all_levels_list)}
}
