% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Funct.R
\name{lp_norm_k_powers}
\alias{lp_norm_k_powers}
\title{PAD Filtering function}
\usage{
lp_norm_k_powers(exp_matrix, p, k)
}
\arguments{
\item{exp_matrix}{matrix including the fit residual of the dataset and the healthy state model.}

\item{p}{integer}

\item{k}{integer}
}
\value{
A numeric vector including the values produced by the function for each sample in the dataset.
}
\description{
A filtering function for mapper that projects $$R$^n$ into $R$
}
\examples{
\dontrun{
lp_norm_k_powers(disease_state,2,1)}

}
