% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{meta_norm_nobs}
\alias{meta_norm_nobs}
\title{meta_norm_nobs}
\usage{
meta_norm_nobs(list_z_scores, list_nobs, column)
}
\arguments{
\item{list_z_scores}{list of z_scores.}

\item{list_nobs}{list of number of observations.}

\item{column}{column}
}
\value{

}
\description{
meta_norm_nobs
}
\examples{
\dontrun{
meta_norm_nobs(list_z_scores,list_nobs,column)
}
}
