% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{one_D_Mapper}
\alias{one_D_Mapper}
\title{one_D_Mapper}
\usage{
one_D_Mapper(
  Ds_for_an,
  filter_function,
  n_int = 10,
  p = 0.3,
  distance_type = "cor",
  clust_type = "hierarchical",
  linkage_type = "single",
  optimal_clust_mode = "standard",
  n_bins_clust = 10
)
}
\arguments{
\item{Ds_for_an}{Disease-state model data for the complete dataset.}

\item{filter_function}{Filtering function output. It is a vector containing the filtering function values for each sample.}

\item{n_int}{Number of intervals used to create the first sample partition based on filtering values.}

\item{p}{Maximum overlap between intervals. Expressed as a fraction from zero to one.}

\item{distance_type}{Type of distance for the clustering procedure. Can be cor or euclidean.}

\item{optimal_clust_mode}{Method to find the optimal number of clusters. Can be the standard method or the silhouette-based method.}

\item{n_bins_clust}{Number of bing to generate the histogram employed by the standard optimal number of cluster finder method.}
}
\value{
Returns a list with different elements.
}
\description{
Wrapping function to carry out the complete process.
}
\examples{
\dontrun{
one_D_Mapper(Ds_for_an,filter_function,n_int = 10,p = 0.3,distance_type = "cor",optimal_clust_mode =  "standard",n_bins_clust = 10)}
}
