% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/An_Functions.R
\name{percent_table}
\alias{percent_table}
\title{percen_table}
\usage{
percent_table(vec_1, vec_2)
}
\arguments{
\item{vec_1}{vector of groups}

\item{vec_2}{vector of categorical variables.}
}
\value{

}
\description{
Computes percentages of categorical variables in groups of samples.
}
\examples{
\dontrun{
create_output_df_unique_node(out_one_D,exp_filt)
}
}
