% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{perform_GSVA_dataset}
\alias{perform_GSVA_dataset}
\title{perform_GSVA_dataset}
\usage{
perform_GSVA_dataset(
  expression,
  surv_Map_Out,
  path_to_gmt_file,
  thr_groups = 20
)
}
\arguments{
\item{expression}{Gene expression matrix.}

\item{surv_Map_Out}{Output from SurvMap analyses.}

\item{path_to_gmt_file}{Pathway to the gtm file containing the pathways atabase.}

\item{thr_groups}{threshold of the number of samples used to filter out small nodes.}
}
\value{

}
\description{
Carries out GSVA for the complete dataset and camputes the averange values by pathways and group of samples.
}
