% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{perform_wilcoxon_each}
\alias{perform_wilcoxon_each}
\title{perform_wilcoxon_each}
\usage{
perform_wilcoxon_each(exp_data, pheno_data, surv_map_res, thr_groups = 20)
}
\arguments{
\item{exp_data}{Gene expression data for the complete dataset.}

\item{pheno_data}{Pheno data with columns pCh_Sample_Name "sample names", pCh_Status "T for disease and NT for non-disease samples"}

\item{surv_map_res}{surv_map_res object}

\item{thr_groups}{Threshold to filter out groups or nodes with less than n samples.}
}
\value{

}
\description{
perform_wilcoxon_each
}
\examples{
\dontrun{
perform_wilcoxon_each(exp_data,pheno_data,surv_map_res,thr_groups = 20)
}
}
