% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{plot_genes_by_groups}
\alias{plot_genes_by_groups}
\title{plot_genes_by_groups}
\usage{
plot_genes_by_groups(exp_data, genes, out_perform_wil)
}
\arguments{
\item{exp_data}{expression data.}

\item{genes}{genes to be plotted}

\item{out_perform_wil}{differential expression analysis results for multiple groups.}

\item{out_dir}{output directory.}
}
\value{

}
\description{
Generate plots of genes by groups.
}
\examples{
\dontrun{
plot_genes_by_groups(exp_data,genes,out_perform_wil)}
}
