% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{plot_GSVA}
\alias{plot_GSVA}
\title{plot_GSVA}
\usage{
plot_GSVA(GSVA_res, row_text_size = 10)
}
\arguments{
\item{GSVA_res}{list containing the output from the perform_GSVA_dataset function.}

\item{row_text_size}{size of the row labels.}
}
\value{

}
\description{
Plots results from GSVA analyses.
}
\examples{
\dontrun{
plot_GSVA(GSVA_res,row_text_size = 10)}
}
