% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{plot_heatmap_data}
\alias{plot_heatmap_data}
\title{plot_heatmap_data}
\usage{
plot_heatmap_data(out_wilc, selected_genes, exp_data, row_text_size = 10)
}
\arguments{
\item{out_wilc}{object returned by the differential gene expresson function.}

\item{selected_genes}{top differentially expressed genes.}

\item{exp_data}{gene expression matrix.}

\item{row_text_size}{size of the row labels.}
}
\value{

}
\description{
plot_heatmap_data
}
\examples{
\dontrun{
plot_heatmap_data(out_wilc,selected_genes,exp_data,row_text_size = 10)}
}
