% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{plot_mean_connectivity}
\alias{plot_mean_connectivity}
\title{plot_mean_connectivity}
\usage{
plot_mean_connectivity(pwer_table, name_ds, cex_val = 0.7)
}
\arguments{
\item{pwer_table}{a single table of powers computed using the compute_power_tables function.}

\item{name_ds}{name of the dataset from which the power table derives.}

\item{cex_val}{cex value to control the size of the numbers printed in the plot.}
}
\value{

}
\description{
Plots the mean connectivity values for the different tested powers.
}
\examples{
\dontrun{
plot_mean_connectivity(pwer_table,name_ds,cex_val = 0.7)
}
}
