% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{plot_multiple_graphs}
\alias{plot_multiple_graphs}
\title{plot_multiple_graphs}
\usage{
plot_multiple_graphs(
  power_tables,
  type = c("fit", "connectivity"),
  file_to_save
)
}
\arguments{
\item{power_tables}{output from the compute_power_tables function.}

\item{type}{fit or connectivity}

\item{file_to_save}{path and file to save the pdf.}
}
\value{

}
\description{
plots scale-free topology fit or mean connectivity against power for multiple datasets using the output from the compute_power_tables function.
}
\examples{
\dontrun{
plot_multiple_graphs(power_tables,type = c("fit","connectivity"),file_to_save)
}
}
