% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{plot_scale_free_fit}
\alias{plot_scale_free_fit}
\title{plot_scale_free_fit}
\usage{
plot_scale_free_fit(pwer_table, name_ds, cex_val = 0.7)
}
\arguments{
\item{pwer_table}{a single table of powers computed using the compute_power_tables function.}

\item{name_ds}{name of the dataset from which the power table derives.}

\item{cex_val}{cex value to control the size of the numbers printed in the plot.}
}
\value{

}
\description{
Plots the scale-free topology fit for the different tested powers.
}
\examples{
\dontrun{
plot_scale_free_fit(pwer_table,name_ds,cex_val = 0.7)
}
}
