% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{prepare_data_stouffer_p_val}
\alias{prepare_data_stouffer_p_val}
\title{prepare_data_stouffer_p_val}
\usage{
prepare_data_stouffer_p_val(p_Data_split, ME_sig_Z, ME_sig_Nobs)
}
\arguments{
\item{p_Data_split}{p_Data_split}

\item{ME_sig_Z}{ME_sig_Z}

\item{ME_sig_Nobs}{ME_sig_Nobs}
}
\value{

}
\description{
prepare_data_stouffer_p_val
}
\examples{
\dontrun{
prepare_data_stouffer_p_val(p_Data_split,ME_sig_Z,ME_sig_Nobs)
}
}
