% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{prepare_diff_meth_res}
\alias{prepare_diff_meth_res}
\title{prepare_diff_meth_res}
\usage{
prepare_diff_meth_res(res_diff, gen_inf, n_genes)
}
\arguments{
\item{res_diff}{output object from perform_wilcoxon_each function.}

\item{gen_inf}{Information about the genes associated to each methylation probe obtained after applying the rewRanges function to a summarized experiment object}

\item{n_genes}{Number of top methylated genes to retrieve for each Node.}
}
\value{

}
\description{
Prepares differential methylation analysis results, and retrieves the top n probes presenting highest up- and -down differential methylation
}
\examples{
\dontrun{
prepare_diff_meth_res(res_diff,gen_inf,n_genes)
}
}
