% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{samples_in_levels}
\alias{samples_in_levels}
\title{samples_in_levels}
\usage{
samples_in_levels(int_data, filter_function)
}
\arguments{
\item{int_data}{Filter function intervals. The output list produced by the get_intervals_One_D function.}

\item{filter_function}{A vector with the filtering function values for each included sample. It is the output of any of the available filtering functions.}
}
\value{
A list with the samples included in each of the levels.
}
\description{
This function returns a list of vectors containing the samples included at each level.
}
\examples{
\dontrun{
samples_in_levels(int_data,filter_function)}
}
