% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_Analysis_Functions.R
\name{select_powers_aux}
\alias{select_powers_aux}
\title{select_powers_aux}
\usage{
select_powers_aux(pow_tab)
}
\arguments{
\item{pow_tab}{power table}
}
\value{

}
\description{
auxiliar function for the select powers function.
}
\examples{
\dontrun{
select_powers_aux(pow_tab)
}
}
