% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{select_top_diff_genes_groups}
\alias{select_top_diff_genes_groups}
\title{select_top_diff_genes_groups}
\usage{
select_top_diff_genes_groups(res_diff_exp, n_genes)
}
\arguments{
\item{res_diff_exp}{res_diff_exp object.}

\item{n_genes}{top differentially expressed genes.}
}
\value{

}
\description{
Select top differentially expressed genes for each group.
}
