% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Select_And_Surv.R
\name{surivival_analysis_multiple_groups}
\alias{surivival_analysis_multiple_groups}
\title{surivival_analysis_multiple_groups}
\usage{
surivival_analysis_multiple_groups(
  pheno_data,
  out_one_D,
  thr_groups = 50,
  ylim_val = c(0.5, 1),
  xlim_val = c(0, 200),
  type = "node"
)
}
\arguments{
\item{pheno_data}{data.frame with phenotype data. Must include a columns called pCh_DFS_E, pCh_DFS_T, and pCh_Status including the event, time to relapse, and the status of the sample T: Disease NT: healthy}

\item{out_one_D}{Mapper output.}

\item{thr_groups}{Minimum number of samples to include a node in the analysis.}

\item{ylim_val}{y-axis limits for survival plot.}

\item{xlim_val}{x-axis limits for survival plot.}

\item{type}{Divide by node or by pam50 in the latter case a column pam50_frma containing the pam50 classification will be requested.}
}
\value{

}
\description{
Survival analysis and plotting of the clustering results. https://www.reneshbedre.com/blog/survival-analysis.html
}
\examples{
\dontrun{
surivival_analysis_multiple_groups(pheno_data,out_one_D,thr_groups = 50,ylim_val = c(0.5, 1),xlim_val =  c(0,200),type = "node")
}
}
