% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvMap.R
\docType{package}
\name{SurvMap}
\alias{SurvMap}
\title{SurvMap: A package for computing mapper-based survival analysis of transcriptomics data.}
\description{
The SurvMap package provides three categories of important functions:
}
\section{SurvMap functions}{

The SurvMap functions ...
}

