% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_Gen_Exp_Enrich.R
\name{write_results}
\alias{write_results}
\title{write_results}
\usage{
write_results(
  formated_res,
  file_name_path = "/media/data/jaume/BreastCancer/Analysis_Mapper_2/Markers_B.xlsx"
)
}
\arguments{
\item{formated_res}{formatted results derived from the format_results function.}

\item{file_name_path}{path and file to save data.}
}
\value{

}
\description{
write_results
}
\examples{
\dontrun{
write_results(formated_res,file_name_path = "/media/data/jaume/BreastCancer/Analysis_Mapper_2/Markers_B.xlsx")
}
}
