% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flat_Construction.R
\name{flatten_normal_tiss}
\alias{flatten_normal_tiss}
\title{F}
\usage{
flatten_normal_tiss(normal_tiss)
}
\arguments{
\item{normal_tiss}{A normal tissue data gene expression matrix.}
}
\value{
A gene expression matrix containing the flattened versión of the vectors.
}
\description{
F
}
\examples{
flatten_normal_tiss(normal_tissue_matrix)
}
