% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{plot_mapper}
\alias{plot_mapper}
\title{plot_mapper}
\usage{
plot_mapper(mapper_list)
}
\arguments{
\item{mapper_list}{A list produced as an output of the one_D_Mapper function.}
}
\value{
Plots an interactive network using the visNetwork function.
}
\description{
This function produces an interactive network plot using the visNetork function.
}
\examples{
\dontrun{
plot_mapper(mapper_list)}
}
